chore: common 패키지 생성 및 도메인/공통 embeddable 객체 분리
