+++ backend/src/main/java/com/codesquad/airbnb/district/DistrictController.java
@@ -1,6 +1,6 @@
 package com.codesquad.airbnb.district;
 
-import com.codesquad.airbnb.domain.Location;
+import com.codesquad.airbnb.common.embeddable.Location;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
