+++ backend/src/main/java/com/codesquad/airbnb/district/DistrictService.java
@@ -1,7 +1,7 @@
 package com.codesquad.airbnb.district;
 
-import com.codesquad.airbnb.domain.Location;
-import com.codesquad.airbnb.domain.ReviewTotal;
+import com.codesquad.airbnb.common.embeddable.Location;
+import com.codesquad.airbnb.common.embeddable.ReviewStat;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -23,7 +23,7 @@ public class DistrictService {
      */
     public List<DistrictResponse> showPopularDistricts(Location location) {
         return districtRepository.getDistrictsWithLocation(
-            location, new ReviewTotal(scoreThreshold, countThreshold));
+            location, new ReviewStat(scoreThreshold, countThreshold));
     }
 
 }
