--- backend/src/main/java/com/codesquad/airbnb/domain/search/PriceRange.java
@@ -1,16 +1,0 @@
-package com.codesquad.airbnb.domain.search;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class PriceRange {
-
-    private Integer min;
-    private Integer max;
-
-    public boolean isNull() {
-        return min == null && max == null;
-    }
-}
