--- backend/src/main/java/com/codesquad/airbnb/domain/search/Radius.java
@@ -1,16 +1,0 @@
-package com.codesquad.airbnb.domain.search;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class Radius {
-
-    private Double horizontal;
-    private Double vertical;
-
-    public boolean isNull() {
-        return horizontal == null && vertical == null;
-    }
-}
