+++ backend/src/main/java/com/codesquad/airbnb/room/RoomRepositoryCustomImpl.java
@@ -5,13 +5,13 @@ import static com.codesquad.airbnb.reservation.QReservation.reservation;
 import static com.codesquad.airbnb.room.entity.QRoom.room;
 import static com.codesquad.airbnb.room.entity.QRoomDetail.roomDetail;
 
-import com.codesquad.airbnb.domain.GuestGroup;
-import com.codesquad.airbnb.domain.Location;
-import com.codesquad.airbnb.domain.StayPeriod;
-import com.codesquad.airbnb.domain.search.Direction;
-import com.codesquad.airbnb.domain.search.PriceRange;
-import com.codesquad.airbnb.domain.search.Radius;
+import com.codesquad.airbnb.common.embeddable.GuestGroup;
+import com.codesquad.airbnb.common.embeddable.Location;
+import com.codesquad.airbnb.common.embeddable.StayPeriod;
+import com.codesquad.airbnb.room.dto.Direction;
 import com.codesquad.airbnb.room.dto.RoomSearCondition;
+import com.codesquad.airbnb.room.dto.RoomSearCondition.PriceRange;
+import com.codesquad.airbnb.room.dto.RoomSearCondition.Radius;
 import com.codesquad.airbnb.room.entity.Room;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQuery;
