+++ backend/src/main/java/com/codesquad/airbnb/room/dto/RoomSearCondition.java
@@ -1,10 +1,8 @@
 package com.codesquad.airbnb.room.dto;
 
-import com.codesquad.airbnb.domain.GuestGroup;
-import com.codesquad.airbnb.domain.Location;
-import com.codesquad.airbnb.domain.StayPeriod;
-import com.codesquad.airbnb.domain.search.PriceRange;
-import com.codesquad.airbnb.domain.search.Radius;
+import com.codesquad.airbnb.common.embeddable.GuestGroup;
+import com.codesquad.airbnb.common.embeddable.Location;
+import com.codesquad.airbnb.common.embeddable.StayPeriod;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
@@ -18,4 +16,27 @@ public class RoomSearCondition {
     private PriceRange priceRange;
     private StayPeriod stayPeriod;
 
+    @Getter
+    @AllArgsConstructor
+    public static class PriceRange {
+
+        private Integer min;
+        private Integer max;
+
+        public boolean isNull() {
+            return min == null && max == null;
+        }
+    }
+
+    @Getter
+    @AllArgsConstructor
+    public static class Radius {
+
+        private Double horizontal;
+        private Double vertical;
+
+        public boolean isNull() {
+            return horizontal == null && vertical == null;
+        }
+    }
 }
