+++ backend/src/main/java/com/codesquad/airbnb/room/entity/RoomDetail.java
@@ -1,9 +1,9 @@
 package com.codesquad.airbnb.room.entity;
 
-import com.codesquad.airbnb.domain.GuestGroup;
-import com.codesquad.airbnb.domain.RoomGroup;
-import com.codesquad.airbnb.domain.RoomOption;
-import com.codesquad.airbnb.domain.StayTime;
+import com.codesquad.airbnb.common.embeddable.GuestGroup;
+import com.codesquad.airbnb.common.embeddable.StayTime;
+import com.codesquad.airbnb.room.entity.embeddable.RoomGroup;
+import com.codesquad.airbnb.room.entity.embeddable.RoomOption;
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
